auto trigger

+++ build-services.sh
@@ -14,17 +14,11 @@ if [ $? -ne 0 ]; then
   exit 1
 fi
 
-# docker build "$SCRIPT_DIR/pipe" -t safe1/pipe
-# if [ $? -ne 0 ]; then
-#   echo "Failed to build docker image safe1/pipe"
-#   exit 1
-# fi
-
-# docker build "$SCRIPT_DIR/auto" -t safe1/auto
-# if [ $? -ne 0 ]; then
-#   echo "Failed to build docker image safe1/auto"
-#   exit 1
-# fi
+docker build "$SCRIPT_DIR/auto" -t smart_home_backend/auto
+if [ $? -ne 0 ]; then
+  echo "Failed to build docker image smart_home/auto"
+  exit 1
+fi
 
 docker rmi $(docker images -f "dangling=true" -q)
